# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_call_service.RejectCallStruct:
    properties:
      callCreator:
        $ref: '#/definitions/types.JID'
      callId:
        type: string
    type: object
  github_com_EvolutionAPI_evolution-go_pkg_call_service.RingCallStruct:
    properties:
      durationSeconds:
        type: integer
      number:
        type: string
    type: object
...
  types.JID:
    properties:
      device:
        format: int32
        type: integer
      integrator:
        format: int32
        type: integer
      rawAgent:
        format: int32
        type: integer
      server:
        type: string
      user:
        type: string
    type: object
...
info:
  contact: {}
  description: Evolution GO - whatsmeow
  title: Evolution GO
  version: "1.0"
paths:
  /call/reject:
    post:
      consumes:
      - application/json
      description: Reject call
      parameters:
      - description: Call data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.RejectCallStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Reject call
      tags:
      - Call
  /call/ring:
    post:
      consumes:
      - application/json
      description: Place an experimental outbound voice call and hang up automatically
      parameters:
      - description: Call data
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/github_com_EvolutionAPI_evolution-go_pkg_call_service.RingCallStruct'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Ring call
      tags:
      - Call
...
swagger: "2.0"